// packages/server-utils/src/integrations/http/types.ts
export interface HttpClientRequest {
  method?: string;
  path?: string;
  host?: string;
  protocol?: string;
  port?: number;
  end(): void;
  getHeader(name: string): string | string[] | number | undefined;
  getHeaders(): Record<string, string | string[] | undefined | number>;
  setHeader(name: string, value: string | string[] | number): void;
  removeHeader(name: string): void;
  prependListener(event: 'response', listener: (res: HttpIncomingMessage) => void): this;
  prependListener(event: string | symbol, listener: (...args: unknown[]) => void): this;
  on(event: string | symbol, listener: (...args: unknown[]) => void): this;
  once(event: string | symbol, listener: (...args: unknown[]) => void): this;
  listenerCount(event: string | symbol): number;
  removeListener(event: string | symbol, listener: (...args: unknown[]) => void): this;
}

/** Minimal interface for http client RequestOptions */
export interface HttpRequestOptions {
  method?: string;
  protocol?: string | null;
  hostname?: string | null;
  host?: string | null;
  port?: string | number | null;
  path?: string | null;
  headers?: Record<string, string | string[] | undefined | number>;
}
// ...
export interface HttpSocket {
  remoteAddress?: string;
  remotePort?: number;
  localAddress?: string;
  localPort?: number;
}

/** Minimal interface for a Node.js http.IncomingMessage */
export interface HttpIncomingMessage {
  statusCode?: number;
  statusMessage?: string;
  httpVersion?: string;
  url?: string;
  method?: string;
  headers: Record<string, string | string[] | undefined>;
  socket?: HttpSocket;
  aborted?: boolean;
  complete?: boolean;
  resume(): void;
  on(event: 'end', listener: () => void): this;
  on(event: string | symbol, listener: (...args: unknown[]) => void): this;
  addListener(event: 'end', listener: () => void): this;
  addListener(event: string | symbol, listener: (...args: unknown[]) => void): this;
  off(event: string | symbol, listener: (...args: unknown[]) => void): this;
  removeListener(event: string | symbol, listener: (...args: unknown[]) => void): this;
}